+++ NFe.Wsdl/NFe.Wsdl.csproj
@@ -59,10 +59,7 @@
     <Compile Include="Autorizacao\NfeAutorizacao3.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Autorizacao\NFeAutorizacao4.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="Autorizacao\NFeAutorizacao4Nativo.cs" />
+    <Compile Include="Autorizacao\NFeAutorizacao4.cs" />
     <Compile Include="Autorizacao\NfeRetAutorizacao3.cs">
       <SubType>Component</SubType>
     </Compile>
